Solve issues after merge from `core-refactor`
