+++ tests/unit/steps/task/test_base.py
@@ -17,12 +17,11 @@ from typing import TYPE_CHECKING, List
 import pytest
 from distilabel.llm.base import LLM
 from distilabel.pipeline.local import Pipeline
-from distilabel.pipeline.step.task.base import Task
-from distilabel.pipeline.step.task.text_generation import TextGeneration
+from distilabel.steps.task.base import Task
 from pydantic import ValidationError
 
 if TYPE_CHECKING:
-    from distilabel.pipeline.step.task.typing import ChatType
+    from distilabel.steps.task.typing import ChatType
 
 
 class DummyLLM(LLM):
@@ -37,11 +36,26 @@ class DummyLLM(LLM):
         return ["output" for _ in inputs]
 
 
-class TestTextGeneration:
+class DummyTask(Task):
+    @property
+    def inputs(self) -> List[str]:
+        return ["instruction"]
+
+    def format_input(self, input: dict) -> "ChatType":
+        return [
+            {"role": "system", "content": ""},
+            {"role": "user", "content": input["instruction"]},
+        ]
+
+    def format_output(self, output: str) -> dict:
+        return {"output": output}
+
+
+class TestTask:
     def test_passing_pipeline(self) -> None:
         pipeline = Pipeline()
         llm = DummyLLM()
-        task = TextGeneration(name="task", llm=llm, pipeline=pipeline)
+        task = DummyTask(name="task", llm=llm, pipeline=pipeline)
         assert task.name == "task"
         assert task.llm is llm
         assert task.generation_kwargs == {}
@@ -50,7 +64,7 @@ class TestTextGeneration:
     def test_within_pipeline_context(self) -> None:
         with Pipeline() as pipeline:
             llm = DummyLLM()
-            task = TextGeneration(name="task", llm=llm, pipeline=pipeline)
+            task = DummyTask(name="task", llm=llm, pipeline=pipeline)
             assert task.name == "task"
             assert task.llm is llm
             assert task.generation_kwargs == {}
@@ -58,12 +72,12 @@ class TestTextGeneration:
 
     def test_with_errors(self) -> None:
         with pytest.raises(ValueError, match="Step 'task' hasn't received a pipeline"):
-            TextGeneration(name="task", llm=DummyLLM())
+            DummyTask(name="task", llm=DummyLLM())
 
         with pytest.raises(
             ValidationError, match="llm\n  Field required \\[type=missing"
         ):
-            TextGeneration(name="task", pipeline=Pipeline())  # type: ignore
+            DummyTask(name="task", pipeline=Pipeline())  # type: ignore
 
         with pytest.raises(
             TypeError,
@@ -74,15 +88,15 @@ class TestTextGeneration:
     def test_process(self) -> None:
         pipeline = Pipeline()
         llm = DummyLLM()
-        task = TextGeneration(name="task", llm=llm, pipeline=pipeline)
+        task = DummyTask(name="task", llm=llm, pipeline=pipeline)
         assert list(task.process([{"instruction": "test"}])) == [
-            [{"instruction": "test", "generation": "output", "model_name": "test"}]
+            [{"instruction": "test", "output": "output", "model_name": "test"}]
         ]
 
     def test_serialization(self) -> None:
         pipeline = Pipeline()
         llm = DummyLLM()
-        task = TextGeneration(name="task", llm=llm, pipeline=pipeline)
+        task = DummyTask(name="task", llm=llm, pipeline=pipeline)
         assert task.dump() == {
             "name": "task",
             "input_mappings": {},
@@ -90,7 +104,7 @@ class TestTextGeneration:
             "input_batch_size": 50,
             "llm": {
                 "type_info": {
-                    "module": "tests.unit.pipeline.step.task.test_text_generation",
+                    "module": "tests.unit.steps.task.test_base",
                     "name": "DummyLLM",
                 }
             },
@@ -103,11 +117,11 @@ class TestTextGeneration:
                 }
             ],
             "type_info": {
-                "module": "distilabel.pipeline.step.task.text_generation",
-                "name": "TextGeneration",
+                "module": "tests.unit.steps.task.test_base",
+                "name": "DummyTask",
             },
         }
 
         with Pipeline() as pipeline:
-            new_task = TextGeneration.from_dict(task.dump())
-            assert isinstance(new_task, TextGeneration)
+            new_task = DummyTask.from_dict(task.dump())
+            assert isinstance(new_task, DummyTask)
